added guns levels and grfx
a lot of 'em

+++ Scripts/Pawns/Pawn_01.gd
@@ -127,7 +127,7 @@ func _process(delta):
 	if nrg != last_nrg:
 		nrg_update()
 		last_nrg = nrg
-	if my_gun:
+	if my_gun != null:
 		my_gun.is_right = is_right
 		my_gun.shoot_pos = shoot_spot
 		
@@ -240,7 +240,7 @@ func jump_rel():
 
 ##-----------------------------------------------------------------------[Shoot]
 func shoot_j():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot_j()
@@ -249,7 +249,7 @@ func shoot_j():
 		my_start_gun.is_right = is_right
 		my_start_gun.shoot_j()
 func shoot():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot()
@@ -258,7 +258,7 @@ func shoot():
 		my_start_gun.is_right = is_right
 		my_start_gun.shoot()
 func shoot_r():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot_r()
@@ -270,16 +270,19 @@ func shoot_r():
 ##-----------------------------------------------------------------------[Throw]
 func pick_throw( left_input, right_input, up_input, down_input, hold_input):
 	if is_holding == true:
-		my_gun.is_right = is_right
-		my_gun.shoot_pos = shoot_spot
+		if my_gun != null:
+			my_gun.is_right = is_right
+			my_gun.shoot_pos = shoot_spot
 		take_ammo = false
 		is_holding = false
 		if !left_input && !right_input && !up_input && !down_input && !hold_input:
 			SFX.play("Blip_11")
-			my_gun.drop()
+			if my_gun != null:
+				my_gun.drop()
 		else:
 			SFX.play("Blip_06")
-			my_gun.throw()
+			if my_gun != null:
+				my_gun.throw()
 		my_gun = null
 		if my_start_gun && start_equiped:
 			my_start_gun.visible = true
@@ -292,8 +295,9 @@ func let_go():
 	if is_holding == true:
 		take_ammo = false
 		is_holding = false
-		my_gun.drop()
-		my_gun = null
+		if my_gun != null:
+			my_gun.drop()
+			my_gun = null
 
 func pick_up():
 	SFX.play("Blip_04")
@@ -317,8 +321,9 @@ func equip_weap(_weap_num, _ammo_pick_up, _time_left, _just_shot):
 	gun_pos.add_child(g)
 	g.init(_ammo_pick_up, player, _time_left, _just_shot)
 	take_ammo = g.take_ammo
-	my_gun = g
-	is_holding = true
+	if g != null:
+		my_gun = g
+		is_holding = true
 
 func equip_start_weap():
 	var g = Equipment.get_weap_hold(0).instance()
@@ -477,7 +482,7 @@ func add_nrg(_nrg):
 
 func add_ammo(_ammo):
 	if take_ammo:
-		if my_gun:
+		if my_gun != null:
 			my_gun.add_ammo(_ammo)
 
 ##-------------------------------------------------------------------[Animation]
@@ -719,7 +724,7 @@ func _set_gun_dir():
 		elif shoot_spot == 6:
 			arm.rotation_degrees = 0
 			arm.bend(3)
-		if my_gun:
+		if my_gun != null:
 			arm.rotation_degrees -= my_gun.walk
 	else:
 		if shoot_spot == 3:
@@ -740,7 +745,7 @@ func _set_gun_dir():
 		elif shoot_spot == 6:
 			arm.rotation_degrees = 0
 			arm.bend(3)
-		if my_gun:
+		if my_gun != null:
 			arm.rotation_degrees += my_gun.walk
 
 func _body(_num: int):
