+++ Scripts/Pawns/Pawn_10.gd
@@ -150,12 +150,12 @@ func _process(delta):
 	if nrg != last_nrg:
 		nrg_update()
 		last_nrg = nrg
-	if my_gun:
+	if my_gun != null:
 		my_gun.is_right = is_right
 		my_gun.shoot_pos = shoot_spot
-	if my_gun:
-		my_gun.is_right = is_right
-		my_gun.shoot_pos = shoot_spot
+#	if my_gun:
+#		my_gun.is_right = is_right
+#		my_gun.shoot_pos = shoot_spot
 		
 	elif start_equiped:
 		my_start_gun.is_right = is_right
@@ -261,7 +261,7 @@ func jump_rel():
 
 ##-----------------------------------------------------------------------[Shoot]
 func shoot_j():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot_j()
@@ -270,7 +270,7 @@ func shoot_j():
 		my_start_gun.is_right = is_right
 		my_start_gun.shoot_j()
 func shoot():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot()
@@ -279,7 +279,7 @@ func shoot():
 		my_start_gun.is_right = is_right
 		my_start_gun.shoot()
 func shoot_r():
-	if my_gun:
+	if my_gun != null:
 		my_gun.shoot_pos = shoot_spot
 		my_gun.is_right = is_right
 		my_gun.shoot_r()
@@ -295,11 +295,11 @@ func pick_throw( left_input, right_input, up_input, down_input, hold_input):
 		is_holding = false
 		if !left_input && !right_input && !up_input && !down_input && !hold_input:
 			SFX.play("Blip_11")
-			if my_gun:
+			if my_gun != null:
 				my_gun.drop()
 		else:
 			SFX.play("Blip_06")
-			if my_gun:
+			if my_gun != null:
 				my_gun.throw()
 		my_gun = null
 		if my_start_gun && start_equiped:
@@ -313,7 +313,8 @@ func let_go():
 	if is_holding == true:
 		take_ammo = false
 		is_holding = false
-		my_gun.drop()
+		if my_gun != null:
+			my_gun.drop()
 		my_gun = null
 
 func pick_up():
@@ -537,7 +538,7 @@ func add_nrg(_nrg):
 
 func add_ammo(_ammo):
 	if take_ammo:
-		if my_gun:
+		if my_gun != null:
 			my_gun.add_ammo(_ammo)
 
 ##-------------------------------------------------------------------[Animation]
@@ -747,7 +748,7 @@ func _set_gun_dir():
 			arm.rotation_degrees = 0
 			head.rotation_degrees = 0
 			arm.bend(3)
-		if my_gun:
+		if my_gun != null:
 			arm.rotation_degrees -= my_gun.walk
 	else:
 		arm.is_right(false)
@@ -776,7 +777,7 @@ func _set_gun_dir():
 			arm.rotation_degrees = 0
 			head.rotation_degrees = 0
 			arm.bend(3)
-		if my_gun:
+		if my_gun != null:
 			arm.rotation_degrees += my_gun.walk
 
 ##-----------------------------------------------------------------------[Color]
